feature/MTCE-249-Monorepo-v6-refactoring-monorepo-directory
https://eonx.atlassian.net/issues/MTCE-249
