+++ monorepo/Command/ExportPackagesAsJsonCommand.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace EonX\EasyMonorepo\Console\Commands;
+namespace EonX\EasyMonorepo\Command;
 
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
