+++ monorepo/Kernel/MonorepoKernel.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace EonX\EasyMonorepo;
+namespace EonX\EasyMonorepo\Kernel;
 
 use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\HttpKernel\Kernel;
@@ -20,6 +20,6 @@ final class MonorepoKernel extends Kernel
 
     public function registerContainerConfiguration(LoaderInterface $loader): void
     {
-        $loader->load(__DIR__ . '/../config/monorepo_services.php');
+        $loader->load(__DIR__ . '/../../config/monorepo_services.php');
     }
 }
