+++ quality/phpstan.neon
@@ -2,7 +2,7 @@ includes:
     - vendor/symplify/phpstan-extensions/config/config.neon
 
 services:
-    - class: EonX\EasyMonorepo\PHPStan\SymfonyMessengerEnvelopeLastReturnType
+    - class: EonX\EasyMonorepo\TypeExtension\SymfonyMessengerEnvelopeLastReturnType
       tags: [phpstan.broker.dynamicMethodReturnTypeExtension]
 
 parameters:
